rename: comment 패키지 위치 이동 및 정리
