+++ src/main/java/com/Sucat/domain/comment/dto/CommentPostResponse.java
@@ -1,14 +1,14 @@
-package com.Sucat.domain.board.DTO;
+package com.Sucat.domain.comment.dto;
 
-public class CommentPostDTO {
+public class CommentPostResponse {
     private String name;
     private String commentContent;
     private String minute;
     private int likeCount;
     private int commentCount;
     private int scrapCount;
-    
-    public CommentPostDTO(String name, String commentContent, String minute, int likeCount, int commentCount, int scrapCount, String imageUrl) {
+
+    public CommentPostResponse(String name, String commentContent, String minute, int likeCount, int commentCount, int scrapCount) {
         this.name = name;
         this.commentContent = commentContent;
         this.minute = minute;
